safari-600.7-branch/LayoutTests/inspector-obsolete/profiler/cpu-profiler-profiling-expected.txt

Tests that CPU profiling works. Bug 52634.

Profiler was enabled.

Running: testProfiling
found pageFunction

Profiler was disabled.
